# [dbnode] Optimize `filesetFiles` function