--- a/src/dbnode/persist/fs/files.go
+++ b/src/dbnode/persist/fs/files.go
@@ -1232,6 +1232,61 @@ func SortedCommitLogFiles(commitLogsDir string) ([]string, error) {
 }
 
 type toSortableFn func(files []string) sort.Interface
+type toBlockStartAndVolumeIndexFn func(file string) (xtime.UnixNano, int, error)
+type sortedFilesetFiles []filesetFile
+
+func (s sortedFilesetFiles) Len() int {
+	return len(s)
+}
+
+func (s sortedFilesetFiles) Less(i, j int) bool {
+	ti := s[i].blockStart
+	tj := s[j].blockStart
+
+	if ti.Before(tj) {
+		return true
+	}
+
+	ij := s[j].volumeIndex
+	ii := s[i].volumeIndex
+	return ti.Equal(tj) && ii < ij
+}
+
+func (s sortedFilesetFiles) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+
+type filesetFile struct {
+	fileName    string
+	blockStart  xtime.UnixNano
+	volumeIndex int
+}
+
+func findFilesEx(fileDir string, pattern string, fn toBlockStartAndVolumeIndexFn) (sortedFilesetFiles, error) {
+	matched, err := filepath.Glob(path.Join(fileDir, pattern))
+	if err != nil {
+		return nil, err
+	}
+	if len(matched) == 0 {
+		return nil, nil
+	}
+	result := make([]filesetFile, len(matched))
+	for i, file := range matched {
+		blockStart, volume, err := fn(file)
+		if err != nil {
+			return nil, err
+		}
+
+		result[i] = filesetFile{
+			fileName:    file,
+			blockStart:  blockStart,
+			volumeIndex: volume,
+		}
+	}
+
+	sort.Sort(sortedFilesetFiles(result))
+	return result, nil
+}
 
 func findFiles(fileDir string, pattern string, fn toSortableFn) ([]string, error) {
 	matched, err := filepath.Glob(path.Join(fileDir, pattern))
@@ -1278,7 +1333,7 @@ type filesetFilesSelector struct {
 
 func filesetFiles(args filesetFilesSelector) (FileSetFilesSlice, error) {
 	var (
-		byTimeAsc []string
+		byTimeAsc sortedFilesetFiles
 		err       error
 	)
 	switch args.fileSetType {
@@ -1286,14 +1341,10 @@ func filesetFiles(args filesetFilesSelector) (FileSetFilesSlice, error) {
 		switch args.contentType {
 		case persist.FileSetDataContentType:
 			dir := ShardDataDirPath(args.filePathPrefix, args.namespace, args.shard)
-			byTimeAsc, err = findFiles(dir, args.pattern, func(files []string) sort.Interface {
-				return dataFileSetFilesByTimeAndVolumeIndexAscending(files)
-			})
+			byTimeAsc, err = findFilesEx(dir, args.pattern, TimeAndVolumeIndexFromDataFileSetFilename)
 		case persist.FileSetIndexContentType:
 			dir := NamespaceIndexDataDirPath(args.filePathPrefix, args.namespace)
-			byTimeAsc, err = findFiles(dir, args.pattern, func(files []string) sort.Interface {
-				return fileSetFilesByTimeAndVolumeIndexAscending(files)
-			})
+			byTimeAsc, err = findFilesEx(dir, args.pattern, TimeAndVolumeIndexFromFileSetFilename)
 		default:
 			return nil, fmt.Errorf("unknown content type: %d", args.contentType)
 		}
@@ -1307,9 +1358,7 @@ func filesetFiles(args filesetFilesSelector) (FileSetFilesSlice, error) {
 		default:
 			return nil, fmt.Errorf("unknown content type: %d", args.contentType)
 		}
-		byTimeAsc, err = findFiles(dir, args.pattern, func(files []string) sort.Interface {
-			return fileSetFilesByTimeAndVolumeIndexAscending(files)
-		})
+		byTimeAsc, err = findFilesEx(dir, args.pattern, TimeAndVolumeIndexFromFileSetFilename)
 	default:
 		return nil, fmt.Errorf("unknown type: %d", args.fileSetType)
 	}
@@ -1328,51 +1377,27 @@ func filesetFiles(args filesetFilesSelector) (FileSetFilesSlice, error) {
 		filesetFiles      = []FileSetFile{}
 	)
 	for _, file := range byTimeAsc {
-		var (
-			currentFileBlockStart xtime.UnixNano
-			volumeIndex           int
-			err                   error
-		)
-		switch args.fileSetType {
-		case persist.FileSetFlushType:
-			switch args.contentType {
-			case persist.FileSetDataContentType:
-				currentFileBlockStart, volumeIndex, err = TimeAndVolumeIndexFromDataFileSetFilename(file)
-			case persist.FileSetIndexContentType:
-				currentFileBlockStart, volumeIndex, err = TimeAndVolumeIndexFromFileSetFilename(file)
-			default:
-				return nil, fmt.Errorf("unknown content type: %d", args.contentType)
-			}
-		case persist.FileSetSnapshotType:
-			currentFileBlockStart, volumeIndex, err = TimeAndVolumeIndexFromFileSetFilename(file)
-		default:
-			return nil, fmt.Errorf("unknown type: %d", args.fileSetType)
-		}
-		if err != nil {
-			return nil, err
-		}
-
 		if latestBlockStart == 0 {
 			latestFileSetFile = NewFileSetFile(FileSetFileIdentifier{
 				Namespace:   args.namespace,
-				BlockStart:  currentFileBlockStart,
+				BlockStart:  file.blockStart,
 				Shard:       args.shard,
-				VolumeIndex: volumeIndex,
+				VolumeIndex: file.volumeIndex,
 			}, args.filePathPrefix)
-		} else if !currentFileBlockStart.Equal(latestBlockStart) || latestVolumeIndex != volumeIndex {
+		} else if !file.blockStart.Equal(latestBlockStart) || latestVolumeIndex != file.volumeIndex {
 			filesetFiles = append(filesetFiles, latestFileSetFile)
 			latestFileSetFile = NewFileSetFile(FileSetFileIdentifier{
 				Namespace:   args.namespace,
-				BlockStart:  currentFileBlockStart,
+				BlockStart:  file.blockStart,
 				Shard:       args.shard,
-				VolumeIndex: volumeIndex,
+				VolumeIndex: file.volumeIndex,
 			}, args.filePathPrefix)
 		}
 
-		latestBlockStart = currentFileBlockStart
-		latestVolumeIndex = volumeIndex
+		latestBlockStart = file.blockStart
+		latestVolumeIndex = file.volumeIndex
 
-		latestFileSetFile.AbsoluteFilePaths = append(latestFileSetFile.AbsoluteFilePaths, file)
+		latestFileSetFile.AbsoluteFilePaths = append(latestFileSetFile.AbsoluteFilePaths, file.fileName)
 	}
 
 	filesetFiles = append(filesetFiles, latestFileSetFile)