@soundvibe 2021-11-08
> Would you care to share the benchmark code (would be enough to put it in PR description)? I am surprised that the improvement is only 30%. What was the sample size for benchmarking?

So initially I was benchmarking with only 100 files so the performance gain was quite low. Tested with more files and most of the time it is ~2x improvement.

```
func BenchmarkFilesetFiles(b *testing.B) {
	shard := uint32(0)
	numIters := 100
	dir := createDataFilesWithVolumeIndex(nil, dataDirName, testNs1ID, shard, numIters, true, dataFileSuffix, 0)
	defer os.RemoveAll(dir)

	for i := 0; i < b.N; i++ {
		files, err := filesetFiles(filesetFilesSelector{
			fileSetType:    persist.FileSetFlushType,
			contentType:    persist.FileSetDataContentType,
			filePathPrefix: dir,
			namespace:      testNs1ID,
			shard:          shard,
			pattern:        filesetFilePrefix + "*",
		})
		if err != nil {
			b.Fatal(err)
		}

		if len(files) == 0 {
			b.Fatal("no files found")
		}
	}
}
```

`-test.benchtime 3s`
// new
numIters := 1000
BenchmarkFilesetFiles-12    	    2124	   1611317 ns/op
BenchmarkFilesetFiles-12    	    2108	   1619529 ns/op
numIters := 10000
BenchmarkFilesetFiles-12    	     121	  27584583 ns/op
BenchmarkFilesetFiles-12    	     116	  26826172 ns/op
numIters := 20000
BenchmarkFilesetFiles-12    	      21	 146918829 ns/op
BenchmarkFilesetFiles-12    	      18	 188116477 ns/op

//old
numIters := 1000
BenchmarkFilesetFiles-12    	    1189	   3033102 ns/op
BenchmarkFilesetFiles-12    	    1174	   2936289 ns/op
numIters := 10000
BenchmarkFilesetFiles-12    	      55	  56059259 ns/op
BenchmarkFilesetFiles-12    	      58	  54876307 ns/op
numIters := 20000
BenchmarkFilesetFiles-12    	      14	 233025020 ns/op
BenchmarkFilesetFiles-12    	      10	 308521944 ns/op