@linasm 2021-11-08
> func BenchmarkFilesetFiles(b *testing.B) {
> 	shard := uint32(0)
> 	numIters := 100
> 	dir := createDataFilesWithVolumeIndex(nil, dataDirName, testNs1ID, shard, numIters, true, dataFileSuffix, 0)
> 	defer os.RemoveAll(dir)

This is benchmarking creation of the files on disk. You need to call `b.ResetTimer()` between benchmark setup and the actual benchmark loop.