@soundvibe 2021-11-08
> > func BenchmarkFilesetFiles(b *testing.B) {
> > shard := uint32(0)
> > numIters := 100
> > dir := createDataFilesWithVolumeIndex(nil, dataDirName, testNs1ID, shard, numIters, true, dataFileSuffix, 0)
> > defer os.RemoveAll(dir)
> 
> This is benchmarking creation of the files on disk. You need to call `b.ResetTimer()` between benchmark setup and the actual benchmark loop.

🤦🏻  Thanks for spotting it. Updated results:
```
numIters := 1000
--new
BenchmarkFilesetFiles-12    	    2260	   1600852 ns/op
BenchmarkFilesetFiles-12    	    2282	   1595209 ns/op
--old
BenchmarkFilesetFiles-12    	    1023	   2945657 ns/op
BenchmarkFilesetFiles-12    	    1142	   3255573 ns/op

numIters := 10000
--new
BenchmarkFilesetFiles-12    	     157	  22408915 ns/op
BenchmarkFilesetFiles-12    	     152	  22152958 ns/op
--old
BenchmarkFilesetFiles-12    	      76	  42065111 ns/op
BenchmarkFilesetFiles-12    	      76	  40777160 ns/op

numIters := 20000
--new
BenchmarkFilesetFiles-12    	      62	  51564310 ns/op
BenchmarkFilesetFiles-12    	      69	  48784302 ns/op
--old
BenchmarkFilesetFiles-12    	      30	 103220878 ns/op
BenchmarkFilesetFiles-12    	      30	 104825039 ns/op
```

Still ~2x improvement.